requirements.txt
amqp
asgiref
async-timeout
billiard
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
decorators
Django
django-celery-beat
django-timezone-field
djangorestframework
idna
kombu
prompt-toolkit
psycopg2-binary==2.9.5
python-crontab
python-dateutil
pytz
redis
requests
six
sqlparse
tzdata
urllib3
vine
wcwidth
